Move cli chess board under `play` subcommand
Prepping for a better engine architecture. The simbelmyne crate will
house both the cli player, the perft tooling, and the engine (accessible
through a UCI interface on stdin/stdout)
