+++ simbelmyne/src/main.rs
@@ -1,171 +1,40 @@
-use anyhow::anyhow;
-use chess::bitboard::Bitboard;
-use chess::board::{Piece, Square, Board};
-use chess::movegen::moves::Move;
-use chess::util::parse;
-use colored::*;
-use engine::Engine;
-use engine::uci::{ClientMessage, ServerMessage};
-use engine::uci::Uci;
-use std::fmt::Display;
-use std::{io, env};
-use std::io::Write;
+use clap::{Parser, Subcommand};
+use play::run_play;
+mod play;
 
-#[cfg(feature = "jemalloc")]
-#[global_allocator]
-static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
-const DEFAULT_FEN: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
-
-struct Game {
-    board: Board,
-    engine: Engine,
-    highlights: Bitboard,
+#[derive(Parser)]
+#[command(author = "Sam Roelants", version = "0.1", about = "A simple perft tool.", long_about = None)]
+struct Cli {
+    #[command(subcommand)]
+    command: Option<Command>,
 }
 
-//TODO: We really don't need a game struct at the moment, just have a board
-// state instead
-impl Game {
-    fn play_move(&mut self, mv: Move) {
-        self.board = self.board.play_move(mv);
-    }
-
-
-    fn get_move(&mut self) -> anyhow::Result<Move> {
-        println!("{self}");
-
-        let selected_square = get_instruction("Move which piece?\n > ")?;
-        let selected_piece = self.try_select(selected_square)?;
-
-        let legal_moves = self.board.legal_moves();
-
-        let mut highlights = Bitboard::EMPTY;
-        highlights |= selected_piece.position;
-        highlights |= legal_moves
-            .iter()
-            .filter(|mv| mv.src() == selected_square)
-            .map(|mv| Bitboard::from(mv.tgt()))
-            .collect();
-
-        self.highlights = highlights;
-
-        println!("{self}");
-
-        let to = get_instruction(&format!(
-            "Move where to?\n {} > ",
-            selected_square.to_alg().bright_blue()
-        ))?;
-
-        self.highlights = Bitboard::EMPTY;
-
-        let mv = legal_moves
-            .into_iter()
-            .filter(|mv| mv.src() == selected_square)
-            .find(|mv| mv.tgt() == to.into())
-            .ok_or(anyhow!("Not a legal move!"))?;
-
-        Ok(mv)
-    }
-
-    fn try_select(&self, square: Square) -> anyhow::Result<&Piece> {
-        let selected = self
-            .board
-            .get_at(square)
-            .ok_or(anyhow!("No piece on square {:?}", square))?;
-
-        if selected.color() != self.board.current {
-            Err(anyhow!("Selected piece belongs to the other player"))?;
-        }
-
-        Ok(selected)
+#[derive(Debug, Subcommand)]
+enum Command {
+    Play {
+        ///Start from a FEN string
+        #[arg(short, long, default_value = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")]
+        fen: String,
     }
 }
 
-pub trait UciClient {
-    fn send(&mut self, msg: ClientMessage) -> anyhow::Result<ServerMessage>;
-}
-
-impl UciClient for Game {
-    fn send(&mut self, msg: ClientMessage) -> anyhow::Result<ServerMessage> {
-        self.engine.receive(msg)
-    }
-}
-
-impl Display for Game {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", "  a b c d e f g h \n".bright_blue())?;
-
-        for rank in (0..8).rev() {
-            write!(f, "{}", (rank + 1).to_string().bright_blue())?;
-            write!(f, " ")?;
-
-            for file in 0..8 {
-                let current_square = Square::new(rank, file);
-
-                let character = self
-                    .board
-                    .get_at(current_square)
-                    .map(|piece| format!("{piece}"))
-                    .unwrap_or(".".to_string());
-
-                if self.highlights.is_empty() || self.highlights.contains(current_square.into()) {
-                    write!(f, "{}", character)?;
-                } else {
-                    write!(f, "{}", character.bright_black())?;
-                }
-
-                write!(f, " ")?;
-            }
-            write!(f, "{}", (rank + 1).to_string().bright_blue())?;
-            write!(f, "\n")?;
-        }
-        write!(f, "{}", "  a b c d e f g h \n".bright_blue())?;
+impl Command {
+    fn run(&self) -> anyhow::Result<()> {
+        match self {
+            Command::Play { fen } => run_play(fen)?,
+        };
 
         Ok(())
     }
 }
 
-fn main() -> anyhow::Result<()>{
-    let fen = env::args().skip(1).next()
-        .unwrap_or(DEFAULT_FEN.to_string());
+fn main() -> anyhow::Result<()> {
+    let cli = Cli::parse();
 
+    if let Some(command) = cli.command {
+        command.run()?;
+    } 
 
-    ////////////////////////////////////////////////////////////////////////////
-    // GAME LOOP
-    ////////////////////////////////////////////////////////////////////////////
-
-    let mut game = Game {
-        board: fen.parse()?,
-        engine: Engine::with_board(fen)?,
-        highlights: Bitboard::EMPTY,
-    };
-
-    loop {
-        println!("FEN: {}", game.board.to_fen());
-        // Get move from the player
-        let mv = game.get_move()?;
-        game.play_move(mv);
-
-        // Update the engine position
-        game.send(ClientMessage::Position(game.board.to_fen()))?;
-
-        // Get move from engine and play it
-        if let ServerMessage::BestMove(mv) = game.send(ClientMessage::Go)? {
-            println!("Got a bessage back: {}", mv);
-            game.play_move(mv);
-        };
-    }
-}
-
-fn get_instruction(prompt: &str) -> anyhow::Result<Square> {
-    let mut input = String::default();
-
-    print!("{prompt}");
-    io::stdout().flush().unwrap();
-    io::stdin().read_line(&mut input).unwrap();
-
-    let (_, square) =
-        parse::algebraic_square(&input).map_err(|_| anyhow!("Invalid square {}", input))?;
-
-    Ok(square)
+    Ok(())
 }
